Separate UserUpdateRequest from the form for the edit user page.
