+++ app/models/Forms.scala
@@ -2,31 +2,26 @@ package models
 
 import play.api.data.{Mapping, Form}
 import play.api.data.Forms._
-import play.api.libs.json.{JsValue, Json}
 
 object Forms {
 
-  case class UserUpdateRequest(
-                                email: String,
-                                username: String,
-                                firstName: Option[String] = None,
-                                lastName: Option[String] = None,
-                                receiveGnmMarketing: Option[Boolean] = None,
-                                receive3rdPartyMarketing: Option[Boolean] = None)
+  case class UserForm(
+    email: String,
+    username: String,
+    firstName: Option[String] = None,
+    lastName: Option[String] = None,
+    receiveGnmMarketing: Option[Boolean] = None,
+    receive3rdPartyMarketing: Option[Boolean] = None) {
 
-  object UserUpdateRequest {
-    def convertUserUpdateRequestToJson(userData: UserUpdateRequest): JsValue = {
-      Json.obj(
-        "email" -> userData.email,
-        "username" -> userData.username,
-        "firstName" -> userData.firstName,
-        "lastName" -> userData.lastName,
-        "receiveGnmMarketing" -> userData.receiveGnmMarketing,
-        "receive3rdPartyMarketing" -> userData.receive3rdPartyMarketing
-      )
-    }
+    val convertToUserUpdateRequest = UserUpdateRequest(
+      this.email,
+      this.username,
+      this.firstName,
+      this.lastName,
+      this.receiveGnmMarketing,
+      this.receive3rdPartyMarketing
+    )
   }
-
 
 //  Experimentation with validation
   private val alphaText: Mapping[String] = text.verifying("First name must be characters only." , name => name.matches("[A-z]+"))
@@ -41,6 +36,6 @@ object Forms {
       "lastName" -> optional(alphaText),
       "receiveGnmMarketing" -> optional(boolean),
       "receive3rdPartyMarketing" -> optional(boolean)
-    )(UserUpdateRequest.apply)(UserUpdateRequest.unapply)
+    )(UserForm.apply)(UserForm.unapply)
   )
 }
