+++ app/views/editUser.scala.html
@@ -1,6 +1,6 @@
 @import play.i18n._
 @import models.Forms._
-@(pageTitle: String, searchQuery: Option[String], user: User,form: Form[UserUpdateRequest], error: Option[String])
+@(pageTitle: String, searchQuery: Option[String], user: User,form: Form[UserForm], error: Option[String])
     @main(pageTitle)(searchQuery){
         <div class="row">
             <div class="col-md-10">
